requirements.txt
fastapi
uvicorn
python-multipart
pillow
opencv-python
numpy
ultralytics
openai==0.28
python-dotenv